+++ CHANGELOG.rst
@@ -2,7 +2,7 @@
  Changelog
 ===========
 
-Unreleased
+4.0.0 - 2017-06-29
 ------------------
 Added
  - Console instances can now be pickled.
@@ -11,7 +11,7 @@ Changed
    and height values.  All returned values will be kept alive by the caller.
 Fixed
  - `Image.get_pixel` now returns a tuple instead of a CData instance.
- - SDL is now initilized lazily.  It should be easier to import BearLibTerminal
+ - SDL is now initialized lazily.  It should be easier to import BearLibTerminal
    along side this library.
 
 3.0.0 - 2017-06-24